\name{basePlot}
\Rdversion{1.0}
\alias{basePlot}
\title{Plot a contour of the 2D Gaussian distribution with covariance matrix K.}
\description{
  Creates the basic plot as an ellipse with major and minor radii as the
  square roots of the two eigenvalues of K.
}
\usage{
  basePlot(K)
}
\arguments{
  \item{K}{the covariance matrix.}
}
\seealso{
  \code{\link{zeroAxes}.}
}
\keyword{model}

