\name{cmpndKernParamInit}
\Rdversion{1.0}
\alias{cmpndKernParamInit}
\title{CMPND kernel parameter initialisation.}
\description{
  initialises the compound kernel structure with some default parameters.
}
\usage{
kern <- cmpndKernParamInit(kern)
}
\arguments{
  \item{kern}{the kernel structure which requires initialisation.}
}
\value{
  \item{kern}{the kernel structure with the default parameters placed in.}
}
\seealso{
\code{\link{kernCreate}, \link{kernParamInit}}.
}
\examples{
## missing
}
\keyword{model}
