\name{cmpndNoiseParamInit}
\Rdversion{1.0}
\alias{cmpndNoiseParamInit}
\title{CMPND noise parameter initialisation.}
\description{
  initialises the compound noise structure with some default parameters.
}
\usage{
cmpndNoiseParamInit(noise,y)
}
\arguments{
  \item{noise}{the noise structure which requires initialisation.}
  \item{y}{the data design matrix.}
}
\value{
  \item{noise}{the noise structure with the default parameters placed in.}
}
\seealso{
\code{\link{noiseCreate}, \link{noiseParamInit}}.
}
\examples{
## missing
}
\keyword{model}
