\name{gaussianNoiseParamInit}
\Rdversion{1.0}
\alias{gaussianNoiseParamInit}
\title{GAUSSIAN noise parameter initialisation.}
\description{
  initialises the Gaussian noise structure with some default parameters.
}
\usage{
noise <- gaussianNoiseParamInit(noise)
}
\arguments{
  \item{noise}{the noise structure which requires initialisation.}
}
\value{
  \item{noise}{the noise structure with the default parameters placed in.}
}
\seealso{
\code{\link{noiseCreate}, \link{noiseParamInit}}.
}
\examples{
## missing
}
\keyword{model}
