\name{gpLogLikelihood}
\Rdversion{1.0}
\alias{gpLogLikelihood}
\title{Compute the log likelihood of a GP.}
\description{
  computes the log likelihood of a data set given a GP model.
}
\usage{
gpLogLikelihood(model)
}
\arguments{
  \item{model}{the GP model for which log likelihood is to be computed.}
}
\value{
  \item{ll}{the log likelihood of the data in the GP model.}
}
\seealso{
\code{\link{gpCreate}, \link{gpLogLikeGradients}, \link{modelLogLikelihood}}.
}
\examples{
## missing
}
\keyword{model}
