\name{gpsim-package}
\alias{gpsim-package}
\alias{gpsim}
\docType{package}
\title{
  gpsim - Gaussian Process Differential Equation Models of Transcription Regulation
}
\description{
  This package implements the method of Gao et al. (2008) and Honkela et
  al. (2010) for Gaussian process modelling single input motif
  regulatory systems with time-series expression data. The method can be
  used to rank potential targets of transcription factors based on such
  data.
}
\details{
  \tabular{ll}{
    Package: \tab gpsim\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2010-01-27\cr
    License: \tab GPL Version 2 or later\cr
  }
  For details of using the package please refer to the Vignette.
}
\author{
  Pei Gao, Antti Honkela, Jonatan Ropponen

  Maintainer: Antti Honkela <antti.honkela@tkk.fi>
}
\references{
  P.~Gao, A.~Honkela, M.~Rattray, and N.~D.~Lawrence.
  Gaussian process modelling of latent chemical species:
  applications to inferring transcription factor activities.
  \emph{Bioinformatics} 24(16):i70--i75, 2008.

  A.~Honkela et al.
  A model-based method for transcription factor target
  identification with limited data.
  Submitted to a journal, 2010.
}
\keyword{ package }
\seealso{
  \code{\link[puma:puma-package]{puma}}
}
\examples{
  # To be added...
}
