\name{kernDisplay}
\Rdversion{1.0}
\alias{kernDisplay}
\title{Display the parameters of the kernel.}
\description{
  Display the parameters of the kernel.
}
\usage{
kernDisplay(kern)
kernDisplay(kern, spacing)
}
\arguments{
  \item{kern}{the kernel to display.}
  \item{spacing}{how many spaces to indent the display of the kernel by.}
}
\details{
  \code{kernDisplay(kern)}
  displays the parameters of the kernel and the kernel type to the console.

  \code{kernDisplay(kern, spacing)}
 