\name{modelDisplay}
\Rdversion{1.0}
\alias{modelDisplay}
\alias{gpdisimDisplay}
\alias{gpsimDisplay}
\alias{cmpndKernDisplay}
\alias{disimKernDisplay}
\alias{kernDisplay}
%\alias{mlpKernDisplay}
\alias{multiKernDisplay}
\alias{rbfKernDisplay}
\alias{simKernDisplay}
%\alias{translateKernDisplay}
\alias{whiteKernDisplay}
\title{Display a model.}
\description{
  displays the parameters of the model/kernel and the model/kernel type to
  the console.
}
\usage{
model <- modelDisplay(model)
}
\arguments{
  \item{model}{the model/kernel structure to be displayed.}
}
\seealso{
\code{\link{modelExtractParam}}
}
\keyword{model}
