\name{modelUpdateProcesses}
\Rdversion{1.0}
\alias{modelUpdateProcesses}
\alias{cgpdisimUpdateProcesses}
\alias{gpdisimUpdateProcesses}
\alias{cgpsimUpdateProcesses}
\alias{gpsimUpdateProcesses}
\title{Update posterior processes of the given model.}
\description{
  updates posterior processes of the given model.
}
\usage{
model <- modelUpdateProcesses(model)
}
\arguments{
  \item{model}{the model whose posterior processes are to be updates.}
}
\value{
  \item{model}{the updated model.}
}
\seealso{
\code{\link{GPPlot}}
}
\examples{
## missing
}
\keyword{model}
