\name{noiseParamInit}
\Rdversion{1.0}
\alias{noiseParamInit}
\title{Noise model's parameter initialisation.}
\description{
  initialises the noise structure with some default parameters.
}
\usage{
noiseParamInit(noise,y)
}
\arguments{
  \item{noise}{the noise structure which requires initialisation.}
  \item{y}{the data design matrix.}
}
\value{
  \item{noise}{the noise structure with the default parameters placed in.}
}
\seealso{
\code{\link{noiseCreate}}.
}
\examples{
## missing
}
\keyword{model}
