\name{zeroAxes}
\Rdversion{1.0}
\alias{zeroAxes}
\title{A function to move the axes crossing point to the origin.}
\description{
  moves the crossing point of the axes to the origin.
}
\usage{
  zeroAxes(col='blue')
}
\arguments{
  \item{col}{color of the axes.}.
}
\seealso{
\code{\link{plot}}.
}
\examples{
## missing
}
\keyword{model}
