\name{gpTest}
\Rdversion{1.0}
\alias{gpTest}
\title{Test the gradients of the gpCovGrads function and the gp models.}
\description{
  runs some tests on the GP-LVM code in the GP toolbox to test that it is working.
}
\usage{
model <- gpTest()
}
\arguments{
}
\value{
  \item{model}{a cell array of models used for testing.}
}
\seealso{
\code{\link{modelTest}}.
}
\examples{
## missing
}
\keyword{model}
