\name{zeroAxes}
\Rdversion{1.0}
\alias{zeroAxes}
\title{A function to move the axes crossing point to the origin.}
\description{
  moves the crossing point of the axes to the origin.
}
\usage{
zeroAxes(axesHandle, tickRatio, fontSize, fontName)
}
\arguments{
  \item{axesHandle}{the handle of the axes to zero.}
  \item{tickRatio}{the ratio of the axis length to the tick length (default 0.025).}
  \item{fontSize}{the font size for the axes (default 14).}
  \item{fontName}{the name of the font for the axes (default 'times'). }
}
\value{
}
\seealso{
\code{\link{plot}}.
}
\examples{
## missing
}
\keyword{model}
